Refactored Spring configuration classes.
